Changing column type of file_ref in sidecar_files to bigint to avoid large numbers being rejected.

+++ src/asset_folder_importer/asset_folder_importer_database.sql
@@ -225,7 +225,7 @@ ALTER SEQUENCE prelude_projects_id_seq OWNED BY prelude_projects.id;
 
 CREATE TABLE sidecar_files (
     id integer NOT NULL,
-    file_ref integer,
+    file_ref bigint,
     sidecar_path text,
     sidecar_name text
 );
